requirements.txt
anyio==4.10.0
APScheduler==3.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-dotenv==1.0.0
python-telegram-bot==21.10
requests==2.32.5
sniffio==1.3.1
tzlocal==5.3.1
urllib3==2.5.0
